+++ include/solvers/navier_stokes_base.h
@@ -910,6 +910,9 @@ protected:
   std::shared_ptr<Quadrature<dim>>     cell_quadrature;
   std::shared_ptr<Quadrature<dim - 1>> face_quadrature;
 
+  // Previous mapping for rotor mesh rotation in mortar method
+  std::shared_ptr<Mapping<dim>> previous_mapping;
+
   // Assemblers for the matrix and rhs
   std::vector<std::shared_ptr<NavierStokesAssemblerBase<dim>>> assemblers;